requirements.txt
django==1.9.2
djangorestframework==3.3.2
sw-django-rest-auth==0.0.4
sw-python-utils==0.0.5
sw-python-sms-devino==0.0.2
factory_boy==2.6.1